Extract methods into Core.Eepi class
