+++ Editor/Ndmf/Core/Eepi.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 95e04c6d5ecef5d47b1f100d8e1a140f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
