+++ Editor/Ndmf/Pass/EepiState.cs
@@ -15,10 +15,19 @@ namespace KusakaFactory.Zatools.Ndmf.Pass
         public ModularAvatarMergeAnimator MergeAnimator => _mergeAnimator;
         public bool Installed => _installed;
 
+        public (Transform Left, Transform Right, Transform Head) TargetAnchors = (null, null, null);
+        public (Transform Left, Transform Right, Transform Head) AnchorProxies = (null, null, null);
+
         private EepiState(BuildContext context)
         {
             _installer = context.AvatarRootObject.GetComponentInChildren<Installer>();
             if (_installer != null) _mergeAnimator = _installer.GetComponent<ModularAvatarMergeAnimator>();
+
+            TargetAnchors = (
+                _installer.transform.Find("WorldAnchor/TargetHand_L"),
+                _installer.transform.Find("WorldAnchor/TargetHand_R"),
+                _installer.transform.Find("WorldAnchor/TargetHead")
+            );
         }
 
         public static EepiState Initializer(BuildContext context) => new EepiState(context);
